Quaternion Camera: Added a quaternion camera
This camera just kind of sits and spins. It is pretty funny looking I guess.
We have to implement the slerp function. There are probably a bunch of errors, but
it behaves the same as glm::slerp, so it probably works for a lot of them. Also, it
pretty blindly follows the definition of a slerp so it should work. It does handle
the case when the sin function would return a 0, resulting in a division by zero.